Separated swayosd into two executables
